Домашнее задание с учетом правок после ревью, спринт 12  Исправлено:  1. Переключатель на "ночной теме" -теперь видно  2. Кнопка "очистить историю" объединила в один контейнер с заголовком "вы искали", теперь не улетает вниз  3. Назвение трека теперь подрезается и видно точку и время

+++ app/src/main/java/com/practicum/playlistmaker/Track.kt
@@ -6,5 +6,5 @@ data class Track (
     val artistName: String,
     val trackTime: String,
     val artworkUrl100: String,
-    val trackTimeMillis: Int
+    val trackTimeMillis: Long
 )
